Added support for calculating average number of questions up to 1020 (above gives infinite values)

+++ headers/ManualSorting.h
@@ -27,7 +27,7 @@ public:
 	void run();
 
 	pair<int, int> getNumOfQuestionsRange(int size);
-	long double getAverageNumOfQuestions(long size);
+	double getAverageNumOfQuestions(int size);
 
 	void readListFromFile();
 	void writeListToFile();
